Add "soversion" to shared library name on Windows to avoid file name conflicts
Summary:
On unix the shared library is named libalkimia[|5].so.7, where '5'
specifies the qt build variant and '7' is the 'soversion', which is not
available on window builds and may conflicts with older installed
versions.

With this patch on Windows the shared library file name is now
libalkimia[|5]7.dll.

Reviewers: #kmymoney, smankowski, tbaumgart

Reviewed By: #kmymoney, tbaumgart

Subscribers: tbaumgart, smankowski

Differential Revision: https://phabricator.kde.org/D10472

(cherry picked from commit b7be7c89e0b0bf23ec71163e0ec99f97f2766644)