Add application icon to online quotes editor for all platforms
Under Windows and Mac, the application icon is embedded into
the application using ecm_add_app_icon(). On Linux this is done
by the installed desktop file.

The icon is based on an SVG icon that has been included.
from the kmymoney sources and modified for use.

The PNG files were created with svg2multipng from the
kdewin-tools project.