Added support for using QtNetwork to download online quotes
To enable the mentioned support, a new cmake variable
BUILD_WITH_QTNETWORK has been added.

When the mentioned cmake variable is set to 1 or ON, which is
default on Windows, AlkOnlineQuotes::launchNative() uses Qt
network support for downloading data. If it is set to 0 or
OFF, the KIO library support is used, which is the default on
non Windows platforms.

The advantage of using Qt network support is less dependency
on 3rd party components, thus less error-proneness and
higher speed on Windows.

BUG:452122
FIXED-IN:8.1.1