Add parsing of online offers in csv format
For this purpose, the "date" and "price" attributes of the online quote
source are used to specify column names from which the information to be
extracted is read, and the "date format" attribute is used to extract the
date.

It is assumed that the first line of the downloaded csv
file contains the names of the columns. The column separator used is
determined automatically, currently supported are [,; \t].

The extracted values are collected in a QMap and sent to the caller via
the signal AlkOnlineQuote::quotes(...).

CCBUG:439411

Signed-off-by: Ralf Habacker <ralf.habacker@freenet.de>