Keep new quote source button enabled
The new quote source button is kept enabled in case the widget is not in
read-only mode. If pressed and an item with name "New Quote Source"
already exists, it is selected and made visible in the combo box.

For Qt5 and above an additional message is displayed, that the item
already exists.

FEATURE: 479773
FIXED-IN: 8.1.73