Fix detecting installed ghns sources
The previous approach required the file name, which was not reliable as
the file name may change between updates.

Now that we have a list of installed online sources from the new material
library, recognition is much more reliable.

As the displayed name of a ghns source is maintained by kde store it can't
be renamed.