More work on Qt6 support
Testing Qt6/KF6 on openSUSE Leap is not so easy at the moment, because the
official KF6 version requires Qt6.5, which is not shipped with Leap.

Furthermore, it is currently not possible to use CI jobs in a private repo
on invent.kde.org: The message reads: This job is stuck because no runners
are assigned to the project online. Go to the project CI settings ...
Project runners - empty - Project runners - These runners are shared by all
projects in this group. ... This project does not belong to a group and
cannot use group runners.