Separate download engines and associated tests
Until now, only Qt or KIO could be selected, which has the disadvantage
that several build configurations are required in order to be able to
compare differences in reaction behaviour. With this commit, both backends
can be activated simultaneously.

In addition, the corresponding test cases have been subdivided more in
order to obtain comparable results.