Use exact match to identify column in CSV download
Yahoo provides the following columns when downloading CSV price
information:

  Date,Open,High,Low,Close,Adj Close,Volume

In case the column to use is named "Close" in the quote source
definition, the information is actually pulled from the column named
"Adj Close" due to the partial match.

This change makes sure to extract the data from the correct column.