Allow to use local script to retrieve quotes
This change allows to provide a local script as URL which will take care
of downloading the data and sends it to stdout.

This change is for Alkimia which now contains the logic to download
price and exchange rate information and fixes the problem in KMyMoney
master. For KMyMoney 5.1.3 a different fix maybe required.

CCBUG: 487477