Add support for reverse online quote search
If a request for the rate of a conversion is not successful, a search with
swapped symbols will be carried out afterwards, if activated, which can
increase the success rate.

As we are about to freeze the user interface, no changes to the user
interface will be introduced with this commit.

BUG:492038
FIXED-IN:8.2