Extend date range support in a URL for downloading quotes
In URLs of online course sources, a start and end date in the
UTC time zone can be used with '%utc' variables in the order
listed. This is necessary for querying prices e.g. at yahoo.com.

For the use of local time zones, the variable '%u' has been
added, which was previously already available when parsing a
date.

CCBUG:495118