Add json file support
The identification of prices and data in a downloaded json file is carried
out in the interface via corresponding values in the input fields 'date'
and 'price' or programmatically via the corresponding methods of the
AlkOnlineQuoteSource class. The hierarchy present in a json file is mapped
by specifying it as a path (with colons as separators).

See autotests/alkonlinequoteprivatetest.cpp for an example.

BUG:495118
FIXED-IN:8.1.90