Avoid not receiving data at the weekend also for sources fetching currencies
To avoid potential problems when retrieving data on weekends, the start
date for currency conversion queries is now also moved to one day before
the weekend. If this happens, a signal is triggered with the changed
date values so that they can be adjusted in the user interface.

Followup for commit e910fe7cb.

BUG:507426
FIXED-IN:5.2.1